feature: attack loop player
The loop checks if any of the entities is already dead.
A message is displayed for how much damage has been inflicted, and
healthPreHit is now inside the loop, as it does not save any resources
to have it outside, due to garbage collection.
Damage is randomized each hit for a more vivid game experience.